[Fix] Scalar: incorret return type from IMAX/IMIN

+++ plugins/scalar/uint/UMESimdVecUint64_2.h
@@ -1079,7 +1079,7 @@ namespace SIMD {
             return mVec[0] > mVec[1] ? 0 : 1;
         }
         // MIMAX
-        UME_FORCE_INLINE uint64_t imax(SIMDVecMask<2> const & mask) const {
+        UME_FORCE_INLINE uint32_t imax(SIMDVecMask<2> const & mask) const {
             uint32_t i0 = 0xFFFFFFFF;
             uint64_t t0 = std::numeric_limits<uint64_t>::min();
             if(mask.mMask[0] == true) {
@@ -1102,11 +1102,11 @@ namespace SIMD {
             return t1;
         }
         // IMIN
-        UME_FORCE_INLINE uint64_t imin() const {
+        UME_FORCE_INLINE uint32_t imin() const {
             return mVec[0] < mVec[1] ? 0 : 1;
         }
         // MIMIN
-        UME_FORCE_INLINE uint64_t imin(SIMDVecMask<2> const & mask) const {
+        UME_FORCE_INLINE uint32_t imin(SIMDVecMask<2> const & mask) const {
             uint32_t i0 = 0xFFFFFFFF;
             uint64_t t0 = std::numeric_limits<uint64_t>::max();
             if(mask.mMask[0] == true) {
